Made ClearableLazy
