+++ src/main/java/net/minecraftforge/common/util/Lazy.java
@@ -24,11 +24,6 @@ public interface Lazy<T> extends Supplier<T>
         return new Lazy.Fast<>(supplier);
     }
 
-    static <T> Lazy<T> ofInvalidatable(@NotNull Supplier<T> supplier)
-    {
-        return new Lazy.InvalidatableFast<>(supplier);
-    }
-
     /**
      * Constructs a thread-safe lazy-initialized object
      * @param supplier The supplier for the value, to be called the first time the value is needed.
@@ -100,31 +95,6 @@ public interface Lazy<T> extends Supplier<T>
                 }
             }
             return instance;
-        }
-    }
-
-    final class InvalidatableFast<T> implements Lazy<T> {
-        private final Supplier<T> supplier;
-        private T instance;
-
-        private InvalidatableFast(Supplier<T> supplier)
-        {
-            this.supplier = supplier;
-        }
-
-        @Nullable
-        @Override
-        public final T get()
-        {
-            if (instance == null)
-            {
-                instance = supplier.get();
-            }
-            return instance;
-        }
-
-        public void invalidate() {
-            this.instance = null;
         }
     }
 }
